Revert "Merge branch 'feature/create-article/S13P31A208-152' into 'be-dev'"
This reverts commit 5b01cfcc0041cfddb29805b8b8b92efe702e3d6b, reversing
changes made to 562119b4d5bb4b868f8e326f6956dec1db5203f9.

+++ backend/src/main/java/com/ssafy/a208/domain/space/service/FolderService.java
@@ -1,15 +1,15 @@
 package com.ssafy.a208.domain.space.service;
 
 
+import com.ssafy.a208.domain.member.entity.Member;
+import com.ssafy.a208.domain.member.reader.MemberReader;
 import com.ssafy.a208.domain.space.dto.request.FolderReq;
 import com.ssafy.a208.domain.space.dto.response.FolderRes;
 import com.ssafy.a208.domain.space.entity.Folder;
 import com.ssafy.a208.domain.space.entity.Space;
-import com.ssafy.a208.domain.space.exception.FolderNotFoundException;
 import com.ssafy.a208.domain.space.reader.FolderReader;
 import com.ssafy.a208.domain.space.repository.FolderRepository;
 import com.ssafy.a208.global.security.dto.CustomUserDetails;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,12 +20,14 @@ public class FolderService {
 
     private final SpaceService spaceService;
     private final FolderReader folderReader;
+    private final MemberReader memberReader;
     private final FolderRepository folderRepository;
 
     @Transactional
     public FolderRes createFolder(CustomUserDetails userDetails, Long spaceId,
             FolderReq folderReq) {
-        Space space = spaceService.getEditableSpace(spaceId, userDetails.memberId());
+        Member member = memberReader.getMemberById(userDetails.memberId());
+        Space space = spaceService.validateAccessToSpace(spaceId, member);
         folderReader.checkDuplicatedName(space, folderReq.name());
 
         Folder folder = Folder.builder()
@@ -46,7 +48,8 @@ public class FolderService {
     @Transactional
     public void updateFolder(CustomUserDetails userDetails, Long spaceId, Long folderId,
             FolderReq folderReq) {
-        Space space = spaceService.getEditableSpace(spaceId, userDetails.memberId());
+        Member member = memberReader.getMemberById(userDetails.memberId());
+        Space space = spaceService.validateAccessToSpace(spaceId, member);
         Folder folder = folderReader.getFolderById(folderId);
         folderReader.checkDuplicatedName(space, folderReq.name());
 
@@ -55,7 +58,8 @@ public class FolderService {
 
     @Transactional
     public void deleteFolder(CustomUserDetails userDetails, Long spaceId, Long folderId) {
-        spaceService.validateEditableSpace(spaceId, userDetails.memberId());
+        Member member = memberReader.getMemberById(userDetails.memberId());
+        spaceService.validateAccessToSpace(spaceId, member);
         Folder folder = folderReader.getFolderById(folderId);
 
         folder.deleteFolder();
@@ -63,7 +67,8 @@ public class FolderService {
 
     @Transactional
     public FolderRes pinFolder(CustomUserDetails userDetails, Long spaceId, Long folderId) {
-        spaceService.validateEditableSpace(spaceId, userDetails.memberId());
+        Member member = memberReader.getMemberById(userDetails.memberId());
+        spaceService.validateAccessToSpace(spaceId, member);
         Folder folder = folderReader.getFolderById(folderId);
 
         folder.updatePin();
@@ -73,17 +78,6 @@ public class FolderService {
                 .color(folder.getColor())
                 .isPinned(folder.isPinned())
                 .build();
-    }
-
-    public Folder getEditableFolder(Long spaceId, Long folderId, Long memberId) {
-        Folder folder = folderReader.getFolderById(folderId);
-        Space space = spaceService.getEditableSpace(spaceId, memberId);
-
-        if (!Objects.equals(folder.getSpace(), space)) {
-            throw new FolderNotFoundException();
-        }
-
-        return folder;
     }
 
 }
